+++ CHANGELOG.md
@@ -4,6 +4,7 @@ Anticipated release: March 23, 2020
 
 #### 🚀 New features
 
+- An all-new activity budget and FFP section ([#2081])
 
 #### 🐛 Bugs fixed
 
@@ -14,3 +15,5 @@ Anticipated release: March 23, 2020
 # Previous releases
 
 See our [release history](https://github.com/18F/cms-hitech-apd/releases)
+
+[#2081]: https://github.com/18F/cms-hitech-apd/pull/2081